#96.4 CAMINO A LA LOCURA

+++ cpu/src/cpu-k-interrupt.h
@@ -4,6 +4,6 @@
 #include "cpu-utils.h"
 int manejar_conexion_kernel_interrupt();
 
-void enviar_proceso_desalojado(int socket_servidor, int pid, int pc);
+void enviar_proceso_desalojado(int socket_servidor, uint8_t pid, uint16_t pc);
 
 #endif // CPU_K_INTERRUPT_H_
